Base new type system
I'm revising the type system using reference material from both JS and C.

The hope is to achieve a new type parsing and object modeling system that can be easily bound to c header syntax. The goal here is not to parse C headers, but to take fragments of them and create accurate retranslations.

It's been said that only shooting stars can break the mould. We shall put that theory to the test.

#Build No. 0.0.605