Added classMethod binding
Oddly enough, class methods seem to require binding. Previously the value of 'this' when a method was invoked, was not the class constructor, but the scoped value of 'this'. It makes not much sense. The methods are defined on an object, so the value of 'this' when the methods are invoked without call, apply, or bind should be the object they are defined upon.

This issue requires some testing for a feasable workaround. As is, binding a method does not provide the method's property interface on the bound property (function). Which means a constructor function would be lost should it require outside manipulation of its prototype using defacto property methods (get, set, delete, etc...) Its an ugly ass hack that we need to get rid of quickly.

Modified: bin/buildnum
Modified: src/JSNative/JSNativeClass.js