Fixed another Error Bug
Perhaps more lie in wait... we shall deal with them in short order.

This fixes a bug where callStack() tries to parse an error from a native code section. InvokeError has wised up due to this fact, and reports the first non-native code section as the source of the error. This is evidently not always going to work out as the source may be in the native code section... However, even native code should have a line number and file. If the error is raised accordingly, I.E. not an 'InvokeError', rather a TypeError or some other Error, this should never be a problem.

Modified: bin/buildnum
Modified: src/JSTools/JSTools.js