Typing Tutor II
JSNative.Type is the "typedef" of JSE. Here we present a simple workable model for dealing with the "standard machine types"

The JSNative.Type constructor should probably make room in arguments to allow for unsigned types and pointers to types. There is no method yet provided to combine the type flags that would be needed to compose such a definiton.

On the other hand, every native machine type should have a proper representation, before we get to "userland" and that representation could be used to obtain a "cast" of the desired flags. This is a threefold addendum as it also mitigates the need to typecheck typecodes, while allowing advanced users and scripts to make or break their own eggs.

Modified: core.js