Officer Friendly: Statutory Arrest
Better declarator parse error 'custody' reporting.

'parse' once again, returns an object, only this time its a pseudo-array-object with a copy of the parsed source as the 'source' member.

I moved the JSONIFY statements outside of the function, as the function was never intended to return or 'echo' a string.

Modified: core.js