Tokenizer gets a boost with seekAhead....
seekAhead works like readAhead, except that the length is obtained by searching  for seekAhead (from current source position) and then subtracting the current source position from that result. The text that will be matched against your rule is all text from current position, up to but not including the text of seekAhead. If your rule's seekAhead marks no content to match, the rule is ignored.