All Wrong
Seperate interface from implementation! That's why it didn't look attractive to me even though it was somewhat effective. I'm moving the project in the direction of Master Pointman as far as the native data interface goes. So what we end up with is basically the addition of JavaScript based classes, and the ability to parse c declarations. The new interface isn't gauranteed to be backwards compatible with pointman, but that's the reason we branched off in the first place. To make improvements. As soon as we have the object classes going, we'll merge back into master and branch off as Squad.

Modified: core.js