Added iStackMachineSlice
Slice creates a fixed subsection of an iStackMachine. Slice and refactor allow this software to be used for much more than value collection and distribution.

Modified: bin/buildnum; synchronized
Modified: main.c; updated example code to reflect additions
Modified: src/JSTools/Stack.inc

Do note that this example relies on the fact that the current implementation returns the content of the first cell when there is no more data to pop.