Non-de-structo
Modified: bin/buildnum

Modified: src/JSTInit.js:
 Added js.extendHidden method for setting non-enumerable object properties

Modified: src/JSTNative.js
 Added native.type.align method which returns the recommended padding for a type given an address base

The native.type.align method is a convenince for calculating struct member alignments.

Right now I am hating the fact that I did not create a Value, Array, and Struct system at the advent of the type system. I'm not bothered by it and nobody is paying me to change it or complaining.... so.... There ye be.

See this great write-up on structure members: http://www.geeksforgeeks.org/structure-member-alignment-padding-and-data-packing/