NOSIX
deprecated posix functions. pay attention here. the plan is to remove all non-inherently compatible functionality, leaving the grunt work to scriptwriters. I noticed that when I write scripts, that's generally what needs to be focused on anyway, the platform you are coding for; and since the end user is best educated on efficacy in this situation, that's where the responsibility lies. its fun to make up a bunch of rules and "stuff" but the best rules aren't really rules, they are merely a framework.