Let run() in scripts have a prototype.
Let the prototype in scripts: run() map the value of stdout, stderr or status toString() (whichever is available first).

Let the prototype in scripts: run() map the status of the command run to valueOf().

Therefore: Calling Number(run(...)) has the effect of status request.

Therefore: Calling String(run(...)) has the effect of stdout, stderr or status; in resolve order.