Using regular values as void * is the best way i can think of to get around platform compatibility issues.
The assumption is that the pointer type uses the largest general register type on th platform. So if we use the largest general exchange type, we can get the largest margin of: NOT TRUNCATED numeric conversions.

In machine types, signedness is only affected by the what we convert the value to in javascript for the most part, when no truncation is involved.