Adding a ghtml Manual testing environment.
Note ghtml is acting funny on my machine, opening in the background. This is a new behavior, and i don't know what causes it.

Normally when this happens i scrap the project, or find a more consistent software offering.

Option 2 will destroy half the project so, buyer beware, ghtml will be stripped of all that fake OOP Goop, becuase 1: its messy as fuck and 2: it causes bullshit to happen as if by wish.

GHTML has only two goals: 1 provide jse with an html platform, 2 allow that platform to be extended with plugins.

If people would build their technology correctly, we wouldn't be troubled by problems like this. all html rendering packages with a script option have their own script engine, instead of plugin api.

* morons in action for ya *