Use largest integer type to write signed/unsigned return value in FunctionCallback
using the long long type to write to the return value (union), will
fill the smallest bits up first, so effectively, all values will fall into
their correct range. Values will not be truncated by this operation if
the correct procedure call format is used for each corresponding type.

If a value is too large (over-read) it will be clipped at its maximum value
if it is unsigned and it will be converted to its minimum value of it is signed.